pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nfl-betting"
version = "0.1.0"
description = "Command-line helpers for NFL betting research using The Odds API and nflfastR data"
readme = "README.md"
authors = [
  { name = "sjb554" }
]
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "requests>=2.31",
  "python-dotenv>=1.0",
  "pydantic>=2.0",
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2"
]

[project.scripts]
nfl-betting = "nfl_betting.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
